Move Assembler into its own module
